2020/4/17

+++ eclassbrand/src/main/java/com/eclass/eclassbrand/DAO/StudentDAO.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.math.BigInteger;
 
 @Repository
-public interface StudentDAO extends JpaRepository<Student, BigInteger> {
+public interface StudentDAO extends JpaRepository<Student, Integer> {
 
     Student findBySno(String sno);
